feat(rust): add two stream examples
